Improve hashing for "rebuild assets" check
This looks at the content directory as well--the reasoning here being
that because we're purging Tailwind, the HTML, etc, in our templates can
have an effect on our styles and thus our assets.

Currently the hash check is not heavily optimized, so for especially
large projects we might need to consider additional work there (I've
also exported that config to a separate file to make it a bit easier to
interact with).