Enable Tailwind JIT processing
The new JIT feature from Tailwind speeds up tailwind processing by a
/lot/, so I've implemented it. This also allows us to reduce the
complexity of the css loading pattern a bit, since we no longer need to
do the weird thing were we try and convince webpack to handle the parts
like separate modules.

Unfortunately (as the message when you run it will tell you) JIT is not
restrained by semver at this time, so it could change or break at any
time. Given that, I'm not going to merge this into main until it hits at
lest some kind of stable release. If you want to play around, though, be
my guest!